Fix refresh gcode preview if already on it when gcode is computed. supermerill/SuperSlicer#2385

+++ src/libslic3r/SLAPrint.hpp
@@ -563,7 +563,7 @@ private:
         void operator()(SLAPrint &         p,
                         double             st,
                         const std::string &msg,
-                        unsigned           flags = SlicingStatus::DEFAULT,
+                        uint16_t           flags = SlicingStatus::DEFAULT,
                         const std::string &logmsg = "");
         
         double status() const { return m_st; }
